requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
langchain-text-splitters
langchain-openai
chromadb
langchain-ollama
faiss-cpu
streamlit